backend/requirements.txt
# ── Core AI ──────────────────────────────────────────────────
google-genai>=1.0.0            # Gemini Live API SDK
google-adk>=0.1.0              # Agent Development Kit

# ── Google Cloud ─────────────────────────────────────────────
google-cloud-storage>=2.14.0   # GCS for document storage
google-cloud-logging>=3.9.0    # Cloud Logging
google-auth>=2.28.0            # Service account auth for Vertex AI
google-cloud-secret-manager>=2.18.0  # Secret Manager for credentials

# ── Backend ──────────────────────────────────────────────────
fastapi>=0.115.0               # WebSocket server
uvicorn[standard]>=0.30.0      # ASGI server
python-multipart>=0.0.9        # File upload support

# ── Persistence ─────────────────────────────────────────────
aiosqlite>=0.19.0              # Async SQLite (local dev)
asyncpg>=0.29.0                # Async PostgreSQL (Cloud SQL production)
aiofiles>=23.0.0               # Async file I/O for uploads

# ── Vision ───────────────────────────────────────────────────
opencv-python-headless>=4.9.0  # Frame processing

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0           # .env file support
reportlab>=4.0.0               # PDF risk report generation
